Setup prettier

+++ server/src/routes/api/messages.ts
@@ -11,7 +11,7 @@ router.get('/', async (req, res) => {
     res.json({
       messages: messages.map((m) => {
         return m.toJSON();
-      }),
+      })
     });
   } catch (err) {
     res.status(500).json({ message: 'Something went wrong.' });
@@ -36,7 +36,7 @@ router.post('/', requireJwtAuth, async (req, res) => {
   try {
     let message = await Message.create({
       text: req.body.text,
-      user: reqUser.id,
+      user: reqUser.id
     });
     message = await message.populate('user').execPopulate();
 
@@ -50,8 +50,7 @@ router.delete('/:id', requireJwtAuth, async (req, res) => {
   try {
     const tempMessage = await Message.findById(req.params.id).populate('user');
     const reqUser: any = req.user;
-    if (!(tempMessage.user.id === reqUser.id || reqUser.role === 'ADMIN'))
-      return res.status(400).json({ message: 'Not the message owner or admin.' });
+    if (!(tempMessage.user.id === reqUser.id || reqUser.role === 'ADMIN')) return res.status(400).json({ message: 'Not the message owner or admin.' });
 
     const message = await Message.findByIdAndRemove(req.params.id).populate('user');
     if (!message) return res.status(404).json({ message: 'No message found.' });
@@ -68,14 +67,9 @@ router.put('/:id', requireJwtAuth, async (req, res) => {
   try {
     const tempMessage = await Message.findById(req.params.id).populate('user');
     const reqUser: any = req.user;
-    if (!(tempMessage.user.id === reqUser.id || reqUser.role === 'ADMIN'))
-      return res.status(400).json({ message: 'Not the message owner or admin.' });
+    if (!(tempMessage.user.id === reqUser.id || reqUser.role === 'ADMIN')) return res.status(400).json({ message: 'Not the message owner or admin.' });
 
-    let message = await Message.findByIdAndUpdate(
-      req.params.id,
-      { text: req.body.text, user: tempMessage.user.id },
-      { new: true },
-    );
+    let message = await Message.findByIdAndUpdate(req.params.id, { text: req.body.text, user: tempMessage.user.id }, { new: true });
     if (!message) return res.status(404).json({ message: 'No message found.' });
     message = await message.populate('user').execPopulate();
